Added base class for OIDCAuthenticator classes
